Create README.md
The project is an Entity classification project which I have implemented using Deep Learning algorithm and TensorFlow framework of Google. I wanted to classify the input entity in 11 categories as First-Name, Last-Name, Designation, Organisation, Website, Email, Mobile, City, State, Pincode, and Country. 
I was unable to find the dataset for the use case so I myself collected a dataset of 12 million records. I passed this dataset to my model to learn and achieved an accuracy of 97.4%.